Mutation and bugfix
Fixed minor bug in the crossover, the dictionary of nodes in the previous version was not update as the crossover happen. The solution used was to walk all the tree and create a new dictionary of nodes. Is not the best solution but is a solution. 

This solution is also used in the mutation, implemented in this version.